requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
defusedxml==0.7.1
fastapi==0.116.1
ffmpeg-python==0.2.0
filelock==3.19.1
fsspec==2025.7.0
future==1.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-genai==1.30.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==1.26.2
openai-whisper==20250625
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
regex==2025.7.34
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
webvtt-py==0.5.1
youtube-transcript-api==1.2.2
yt-dlp==2025.8.11
